refactor: separate security beans and filter chain configs
